add: モックデータ作成関数の切り出し
